refactor(run-protocol): move interest.js up from vaultFactory to be used by getRUN
