+++ packages/run-protocol/test/test-interest.js
@@ -13,7 +13,7 @@ import {
   chargeInterest,
   makeInterestCalculator,
   SECONDS_PER_YEAR,
-} from '../../src/vaultFactory/interest.js';
+} from '../src/interest.js';
 
 const ONE_DAY = 60n * 60n * 24n;
 const ONE_MONTH = ONE_DAY * 30n;
